requirements.txt
# ── Core Framework ────────────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
streamlit==1.35.0
pydantic==2.7.1
pydantic-settings==2.2.1

# ── Agentic / LLM ─────────────────────────────────────────────────────────────
langchain==0.2.5
langchain-community==0.2.5
langchain-groq==0.1.5
langgraph==0.1.14
langchain-huggingface==0.0.3
groq==0.9.0

# ── RAG / Vector Store ────────────────────────────────────────────────────────
faiss-cpu==1.8.0
sentence-transformers==3.0.1
chromadb==0.5.3

# ── Database ──────────────────────────────────────────────────────────────────
sqlalchemy==2.0.30
alembic==1.13.1

# ── HTTP / APIs ───────────────────────────────────────────────────────────────
httpx==0.27.0
tweepy==4.14.0
requests==2.32.3
python-dotenv==1.0.1

# ── Utilities ─────────────────────────────────────────────────────────────────
tenacity==8.3.0
structlog==24.2.0
rich==13.7.1
python-dateutil==2.9.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest==8.2.2
pytest-asyncio==0.23.7
pytest-cov==5.0.0
httpx==0.27.0

# ── Dev / Linting ─────────────────────────────────────────────────────────────
black==24.4.2
isort==5.13.2
mypy==1.10.0